Refactor CLI module and add option builders
